requirements.txt
python-dotenv
discord.py
flask
pytz